# Backend/api-gateway/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Trackify Platform API
  description: |
    Gateway-facing REST API for the Trackify multi-tenant SaaS platform.
    Authenticated routes require a Bearer JWT obtained from `POST /auth/login`.
    All responses may include `X-Correlation-Id` for request tracing.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local API gateway
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    IssueRequest:
      type: object
      required: [title, projectId]
      properties:
        title:
          type: string
        description:
          type: string
        projectId:
          type: integer
          format: int64
        assigneeId:
          type: integer
          format: int64
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        status:
          type: string
          enum: [TODO, IN_PROGRESS, DONE]
        sprintId:
          type: integer
          format: int64
paths:
...
  /tenants:
    get:
      tags: [Tenants]
      summary: List tenants (MASTER)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tenant list
    post:
      tags: [Tenants]
      summary: Create tenant (MASTER)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Tenant created
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects for current tenant
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Project list
    post:
      tags: [Projects]
      summary: Create project
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Project created
...
  /dashboard:
    get:
      tags: [Projects]
      summary: Tenant dashboard widgets (summary, activity, assigned issues, recent projects)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dashboard payload
...
  /issues:
    post:
      tags: [Issues]
      summary: Create issue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueRequest'
      responses:
        '201':
          description: Issue created
...
  /search:
    get:
      tags: [Search]
      summary: Global search across projects, issues, and users
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
  /notifications:
    get:
      tags: [Notifications]
      summary: List in-app notifications
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notifications
...